Swap round PropertyCalculator and PropertyAssessor
