+++ spec/services/collators/capital_collator_spec.rb
@@ -42,10 +42,10 @@ module Collators
         end
 
         it "instantiates and calls the Property Assessment service" do
-          property_result = Calculators::PropertyCalculator::Result.new(assessed_equity: 23_000.0,
+          property_result = Assessors::PropertyAssessor::Result.new(assessed_equity: 23_000.0,
                                                                         property: assessment.applicant_capital_summary.main_home,
                                                                         smod_allowance: 0)
-          allow(Calculators::PropertyCalculator).to receive(:call).and_return([property_result])
+          allow(Assessors::PropertyAssessor).to receive(:call).and_return([property_result])
           expect(collator.total_property).to eq 23_000.0
         end
       end
@@ -141,11 +141,11 @@ module Collators
         end
 
         it "summarizes the results it gets from the subservices" do
-          property_result = Calculators::PropertyCalculator::Result.new(assessed_equity: 23_000.0,
+          property_result = Assessors::PropertyAssessor::Result.new(assessed_equity: 23_000.0,
                                                                         property: assessment.applicant_capital_summary.main_home,
                                                                         smod_allowance: 0)
 
-          allow(Calculators::PropertyCalculator).to receive(:call).and_return([property_result])
+          allow(Assessors::PropertyAssessor).to receive(:call).and_return([property_result])
 
           expect(collator.total_liquid.to_f).to eq 145.83
           expect(collator.total_non_liquid).to eq 500
