+++ spec/services/decorators/v6/property_decorator_spec.rb
@@ -24,7 +24,7 @@ module Decorators
                    shared_with_housing_assoc: false
           end
           let(:property) do
-            result = Calculators::PropertyCalculator.call(submission_date: Date.current, properties: [record],
+            result = Assessors::PropertyAssessor.call(submission_date: Date.current, properties: [record],
                                                           level_of_help: "certificated", smod_cap: 100_000)
             PropertySubtotals.new result.first
           end
